+++ CHANGELOG.rst
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+----------
+Unreleased
+----------
+Added
+-----
+- Time zone support with ``--tz`` option
+
 ---------------------
 `0.5.0`_ - 2021-06-11
 ---------------------