+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
 ----------
-Unreleased
+`0.6.0`_ - 2021-06-20
 ----------
 Added
 -----
@@ -91,3 +91,5 @@ Fixed
 .. _0.3.0: https://github.com/mkdryden/telegram-stats-bot/releases/tag/v0.3.0
 .. _0.3.1: https://github.com/mkdryden/telegram-stats-bot/releases/tag/v0.3.1
 .. _0.4.0: https://github.com/mkdryden/telegram-stats-bot/releases/tag/v0.4.0
+.. _0.5.0: https://github.com/mkdryden/telegram-stats-bot/releases/tag/v0.5.0
+.. _0.6.0: https://github.com/mkdryden/telegram-stats-bot/releases/tag/v0.6.0